+++ tests/ini.c
@@ -30,42 +30,42 @@ static void __iniOnError(CwtIniHandler ini, const CWT_CHAR *errstr)
 	printf_error(_T("%s at line %lu"), errstr, cwtIniNS()->line(ini));
 }
 
-static BOOL __onIniDirective1(CwtIniHandler h, const CWT_CHAR* argv[], int argc)
+static BOOL __onIniDirective1(CwtIniHandler h, const CWT_CHAR* argv[], size_t argc)
 {
 	CWT_TEST_OK(argc == 6);
 	CWT_TEST_OK(cwtStrNS()->streq(_T("directive1"), argv[0]));
 	return TRUE;
 }
 
-static BOOL __onIniDirective2(CwtIniHandler h, const CWT_CHAR* argv[], int argc)
+static BOOL __onIniDirective2(CwtIniHandler h, const CWT_CHAR* argv[], size_t argc)
 {
 	CWT_TEST_OK(argc == 5);
 	CWT_TEST_OK(cwtStrNS()->streq(_T("directive2"), argv[0]));
 	return TRUE;
 }
 
-static BOOL __onIniDirective3(CwtIniHandler h, const CWT_CHAR* argv[], int argc)
+static BOOL __onIniDirective3(CwtIniHandler h, const CWT_CHAR* argv[], size_t argc)
 {
 	CWT_TEST_OK(argc == 4);
 	CWT_TEST_OK(cwtStrNS()->streq(_T("directive3"), argv[0]));
 	return TRUE;
 }
 
-static BOOL __onIniDirective4(CwtIniHandler h, const CWT_CHAR* argv[], int argc)
+static BOOL __onIniDirective4(CwtIniHandler h, const CWT_CHAR* argv[], size_t argc)
 {
 	CWT_TEST_OK(argc == 3);
 	CWT_TEST_OK(cwtStrNS()->streq(_T("directive4"), argv[0]));
 	return TRUE;
 }
 
-static BOOL __onIniDirective5(CwtIniHandler h, const CWT_CHAR* argv[], int argc)
+static BOOL __onIniDirective5(CwtIniHandler h, const CWT_CHAR* argv[], size_t argc)
 {
 	CWT_TEST_OK(argc == 2);
 	CWT_TEST_OK(cwtStrNS()->streq(_T("directive5"), argv[0]));
 	return TRUE;
 }
 
-static BOOL __onIniDirective6(CwtIniHandler h, const CWT_CHAR* argv[], int argc)
+static BOOL __onIniDirective6(CwtIniHandler h, const CWT_CHAR* argv[], size_t argc)
 {
 	CWT_TEST_OK(argc == 1);
 	CWT_TEST_OK(cwtStrNS()->streq(_T("directive6"), argv[0]));
